[#1940] refactored unit tests for view functions
