+++ hs_core/tests/api/views/test_add_delete_resource_files.py
@@ -1,21 +1,16 @@
 import os
-import tempfile
-import shutil
 
-from django.contrib.sessions.middleware import SessionMiddleware
-from django.test import TestCase, RequestFactory
 from django.contrib.auth.models import Group
-from django.contrib.messages.storage.fallback import FallbackStorage
 from django.core.urlresolvers import reverse
 
 from rest_framework import status
 
 from hs_core import hydroshare
 from hs_core.views import add_files_to_resource, delete_file, delete_multiple_files
-from hs_core.testing import MockIRODSTestCaseMixin
+from hs_core.testing import MockIRODSTestCaseMixin, ViewTestCase
 
 
-class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
+class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, ViewTestCase):
     def setUp(self):
         super(TestAddDeleteResourceFiles, self).setUp()
         self.group, _ = Group.objects.get_or_create(name='Hydroshare Author')
@@ -36,9 +31,6 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
             title='Generic Resource Key/Value Metadata Testing'
         )
 
-        self.factory = RequestFactory()
-        self.temp_dir = tempfile.mkdtemp()
-
         # Make a text file
         self.txt_file_name_1 = 'text-1.txt'
         self.txt_file_path_1 = os.path.join(self.temp_dir, self.txt_file_name_1)
@@ -51,11 +43,6 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
         txt = open(self.txt_file_path_2, 'w')
         txt.write("Hello World-2\n")
         txt.close()
-
-    def tearDown(self):
-        super(TestAddDeleteResourceFiles, self).tearDown()
-        if os.path.exists(self.temp_dir):
-            shutil.rmtree(self.temp_dir)
 
     def test_add_files(self):
         # here we are testing add_files_to_resource view function
@@ -72,7 +59,7 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
         request.user = self.user
         # make it a ajax request
         request.META['HTTP_X_REQUESTED_WITH'] = 'XMLHttpRequest'
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         response = add_files_to_resource(request, shortkey=self.gen_res.short_id)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         # there should be 1 file
@@ -92,7 +79,7 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
         request.user = self.user
         # make it a ajax request
         request.META['HTTP_X_REQUESTED_WITH'] = 'XMLHttpRequest'
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         response = add_files_to_resource(request, shortkey=self.gen_res.short_id)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         # there should be 1 file
@@ -114,7 +101,7 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
         request.user = self.user
         # make it a ajax request
         request.META['HTTP_X_REQUESTED_WITH'] = 'XMLHttpRequest'
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         add_files_to_resource(request, shortkey=self.gen_res.short_id)
         res_file = self.gen_res.files.first()
 
@@ -126,8 +113,8 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
         request = self.factory.post(url, data={})
         request.user = self.user
         request.META['HTTP_REFERER'] = 'some-url'
-        self._set_request_message_attributes(request)
-        self._add_session_to_request(request)
+        self.set_request_message_attributes(request)
+        self.add_session_to_request(request)
         response = delete_file(request, shortkey=self.gen_res.short_id, f=res_file.id)
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
         self.assertEqual(response['Location'], request.META['HTTP_REFERER'])
@@ -151,7 +138,7 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
         request.user = self.user
         # make it a ajax request
         request.META['HTTP_X_REQUESTED_WITH'] = 'XMLHttpRequest'
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         add_files_to_resource(request, shortkey=self.gen_res.short_id)
         self.assertEqual(self.gen_res.files.count(), 2)
 
@@ -163,24 +150,11 @@ class TestAddDeleteResourceFiles(MockIRODSTestCaseMixin, TestCase):
         request = self.factory.post(url, data=post_data)
         request.user = self.user
         request.META['HTTP_REFERER'] = 'some-url'
-        self._set_request_message_attributes(request)
-        self._add_session_to_request(request)
+        self.set_request_message_attributes(request)
+        self.add_session_to_request(request)
         response = delete_multiple_files(request, shortkey=self.gen_res.short_id)
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
         self.assertEqual(response['Location'], request.META['HTTP_REFERER'])
         self.assertEqual(self.gen_res.files.count(), 0)
 
         hydroshare.delete_resource(self.gen_res.short_id)
-
-    def _set_request_message_attributes(self, request):
-        # the following 3 lines are for preventing error in unit test due to the view being
-        # tested uses messaging middleware
-        setattr(request, 'session', 'session')
-        messages = FallbackStorage(request)
-        setattr(request, '_messages', messages)
-
-    def _add_session_to_request(self, request):
-        """Annotate a request object with a session"""
-        middleware = SessionMiddleware()
-        middleware.process_request(request)
-        request.session.save()
