+++ hs_core/tests/api/views/test_delete_resource.py
@@ -1,9 +1,6 @@
 import json
 
-from django.contrib.sessions.middleware import SessionMiddleware
-from django.test import TestCase, RequestFactory
 from django.contrib.auth.models import Group
-from django.contrib.messages.storage.fallback import FallbackStorage
 from django.core.urlresolvers import reverse
 
 from rest_framework import status
@@ -11,10 +8,10 @@ from rest_framework import status
 from hs_core import hydroshare
 from hs_core.models import BaseResource
 from hs_core.views import delete_resource
-from hs_core.testing import MockIRODSTestCaseMixin
+from hs_core.testing import MockIRODSTestCaseMixin, ViewTestCase
 
 
-class TestCRUDMetadata(MockIRODSTestCaseMixin, TestCase):
+class TestCRUDMetadata(MockIRODSTestCaseMixin, ViewTestCase):
     def setUp(self):
         super(TestCRUDMetadata, self).setUp()
         self.group, _ = Group.objects.get_or_create(name='Hydroshare Author')
@@ -35,8 +32,6 @@ class TestCRUDMetadata(MockIRODSTestCaseMixin, TestCase):
             title='Generic Resource Key/Value Metadata Testing'
         )
 
-        self.factory = RequestFactory()
-
     def test_delete_resource(self):
         # here we are testing the delete_resource view function
 
@@ -47,23 +42,10 @@ class TestCRUDMetadata(MockIRODSTestCaseMixin, TestCase):
         request.user = self.user
         # make it a ajax request
         request.META['HTTP_X_REQUESTED_WITH'] = 'XMLHttpRequest'
-        self._set_request_message_attributes(request)
-        self._add_session_to_request(request)
+        self.set_request_message_attributes(request)
+        self.add_session_to_request(request)
         response = delete_resource(request, shortkey=self.gen_res.short_id)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         response_dict = json.loads(response.content)
         self.assertEqual(response_dict['status'], 'success')
         self.assertEqual(BaseResource.objects.count(), 0)
-
-    def _set_request_message_attributes(self, request):
-        # the following 3 lines are for preventing error in unit test due to the view being
-        # tested uses messaging middleware
-        setattr(request, 'session', 'session')
-        messages = FallbackStorage(request)
-        setattr(request, '_messages', messages)
-
-    def _add_session_to_request(self, request):
-        """Annotate a request object with a session"""
-        middleware = SessionMiddleware()
-        middleware.process_request(request)
-        request.session.save()
