+++ hs_core/tests/api/views/test_group.py
@@ -1,10 +1,9 @@
 import json
 from mock import patch
 
-from django.test import TestCase, RequestFactory, Client
+from django.test import Client
 from django.core.urlresolvers import reverse
 from django.contrib.auth.models import Group
-from django.contrib.messages.storage.fallback import FallbackStorage
 from django.db import transaction
 from django.utils.http import int_to_base36
 
@@ -16,11 +15,11 @@ from hs_core import hydroshare
 from hs_core.views import create_user_group, update_user_group, share_group_with_user, unshare_group_with_user, \
     make_group_membership_request, act_on_group_membership_request, share_resource_with_group, \
     unshare_resource_with_group, delete_user_group, restore_user_group
-from hs_core.testing import MockIRODSTestCaseMixin
+from hs_core.testing import MockIRODSTestCaseMixin, ViewTestCase
 from hs_access_control.models import PrivilegeCodes
 
 
-class TestGroup(MockIRODSTestCaseMixin, TestCase):
+class TestGroup(MockIRODSTestCaseMixin, ViewTestCase):
 
     def setUp(self):
         super(TestGroup, self).setUp()
@@ -55,7 +54,6 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                                                    title='Test Resource',
                                                    metadata=[]
                                                    )
-        self.factory = RequestFactory()
 
     def test_create_group(self):
         # TODO: test with picture file upload for the group
@@ -64,7 +62,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'name': 'Test Group-1', 'description': 'This is a cool group-1', 'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         response = create_user_group(request)
         new_group = Group.objects.filter(name='Test Group-1').first()
@@ -78,7 +76,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         # test passing privacy_level = 'private'
         grp_data = {'name': 'Test Group-2', 'description': 'This is a cool group-2', 'privacy_level': 'private'}
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         create_user_group(request)
@@ -91,7 +89,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         # test passing privacy_level = 'discoverable'
         grp_data = {'name': 'Test Group-3', 'description': 'This is a cool group-3', 'privacy_level': 'discoverable'}
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         create_user_group(request)
@@ -114,7 +112,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'description': 'This is a cool group', 'privacy_level': 'public'}
 
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
@@ -130,7 +128,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'name': 'Test Group', 'privacy_level': 'public'}
 
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
@@ -146,7 +144,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'name': 'Test Group', 'description': 'This is a cool group'}
 
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
@@ -162,7 +160,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'name': 'Test Group', 'description': 'This is a cool group', 'privacy_level': 'some-level'}
 
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
@@ -177,7 +175,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         # test that duplicate group names are not allowed
         grp_data = {'name': 'Test Group', 'description': 'This is a cool group', 'privacy_level': 'private'}
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         response = create_user_group(request)
@@ -188,7 +186,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         # create a group with duplicate name
         grp_data = {'name': 'Test Group', 'description': 'This is a very cool group', 'privacy_level': 'private'}
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
 
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
@@ -210,7 +208,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'purpose': 'This group has no purpose', 'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         response = create_user_group(request)
         new_group = Group.objects.filter(name='Test Group-1').first()
@@ -233,7 +231,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'purpose': 'This group now has purpose', 'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=new_group.id)
@@ -253,7 +251,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=new_group.id)
@@ -273,7 +271,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'privacy_level': 'private'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=new_group.id)
@@ -293,7 +291,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=new_group.id)
@@ -314,7 +312,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'privacy_level': 'discoverable'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=new_group.id)
@@ -338,7 +336,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'purpose': 'This group has no purpose', 'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         create_user_group(request)
         new_group = Group.objects.filter(name='Test Group-1').first()
@@ -348,7 +346,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url = reverse('delete_user_group', kwargs=post_data)
         request = self.factory.post(url, data=post_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = delete_user_group(request, group_id=new_group.id)
@@ -362,7 +360,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url = reverse('restore_user_group', kwargs=post_data)
         request = self.factory.post(url, data=post_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = restore_user_group(request, group_id=new_group.id)
@@ -382,7 +380,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
                     'purpose': 'This group has purpose', 'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         response = create_user_group(request)
         new_group = Group.objects.filter(name='Test Group-1').first()
@@ -394,7 +392,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'description': 'This is a cool group-2', 'purpose': 'This group has purpose'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=new_group.id)
@@ -422,7 +420,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'name': 'Hydroshare Author', 'description': 'This is a cool group-1',
                     'purpose': 'This group has purpose'}
         request = self.factory.post(url, data=grp_data)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=new_group.id)
@@ -463,7 +461,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url = reverse('share_group_with_user', kwargs=url_params)
         request = self.factory.post(url)
         request.META['HTTP_REFERER'] = "/some_url/"
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         response = share_group_with_user(request, group_id=new_group.id, user_id=self.mike.id, privilege="badprivilege")
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
@@ -485,7 +483,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url = reverse('unshare_group_with_user', kwargs=url_params)
         request = self.factory.post(url)
         request.META['HTTP_REFERER'] = "/some_url/"
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         response = unshare_group_with_user(request, group_id=new_group.id, user_id=self.mike.id)
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
@@ -581,7 +579,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url_params = {'group_id': new_group.id}
         url = reverse('make_group_membership_request', kwargs=url_params)
         request = self.factory.post(url)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.META['HTTP_REFERER'] = "/some_url/"
         request.user = self.mike
         response = make_group_membership_request(request, group_id=new_group.id)
@@ -609,7 +607,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url_params = {'group_id': new_group.id, 'user_id': self.mike.id}
         url = reverse('make_group_membership_request', kwargs=url_params)
         request = self.factory.post(url)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.META['HTTP_REFERER'] = "/some_url/"
         request.user = self.john
         response = make_group_membership_request(request, group_id=new_group.id, user_id=self.mike.id)
@@ -741,7 +739,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
 
     def _update_failure(self, group, request):
         group_name = group.name
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = update_user_group(request, group_id=group.id)
@@ -758,7 +756,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url_params = {'shortkey': self.resource.short_id, 'privilege': privilege, 'group_id': group.id}
         url = reverse('share_resource_with_group', kwargs=url_params)
         request = self.factory.post(url)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         response = share_resource_with_group(request, shortkey=self.resource.short_id, privilege=privilege,
                                              group_id=group.id)
@@ -768,7 +766,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url_params = {'membership_request_id': membership_request.id, 'action': action}
         url = reverse('act_on_group_membership_request', kwargs=url_params)
         request = self.factory.post(url)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         request.META['HTTP_REFERER'] = "/some_url/"
         response = act_on_group_membership_request(request, membership_request_id=membership_request.id,
@@ -781,7 +779,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url_params = {'membership_request_id': membership_request.id, 'action': action}
         url = reverse('act_on_group_membership_request', kwargs=url_params)
         request = self.factory.post(url)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.mike
         request.META['HTTP_REFERER'] = "/some_url/"
         response = act_on_group_membership_request(request, membership_request_id=membership_request.id,
@@ -794,7 +792,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url_params = {'group_id': group.id}
         url = reverse('make_group_membership_request', kwargs=url_params)
         request = self.factory.post(url)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.META['HTTP_REFERER'] = "/some_url/"
         request.user = self.mike
         make_group_membership_request(request, group_id=group.id)
@@ -805,7 +803,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url_params = {'group_id': group.id, 'user_id': self.mike.id}
         url = reverse('make_group_membership_request', kwargs=url_params)
         request = self.factory.post(url)
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.META['HTTP_REFERER'] = "/some_url/"
         request.user = self.john
         make_group_membership_request(request, group_id=group.id, user_id=self.mike.id)
@@ -818,7 +816,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         grp_data = {'name': 'Test Group', 'description': 'This is a cool group', 'privacy_level': 'public'}
         request = self.factory.post(url, data=grp_data)
 
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         create_user_group(request)
         new_group = Group.objects.filter(name='Test Group').first()
@@ -829,7 +827,7 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
         url = reverse('share_group_with_user', kwargs=url_params)
         request = self.factory.post(url)
         request.META['HTTP_REFERER'] = "/some_url/"
-        self._set_request_message_attributes(request)
+        self.set_request_message_attributes(request)
         request.user = self.john
         response = share_group_with_user(request, group_id=group.id, user_id=self.mike.id, privilege=privilege)
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
@@ -844,10 +842,3 @@ class TestGroup(MockIRODSTestCaseMixin, TestCase):
             self.assertIn(self.mike, group.gaccess.get_users_with_explicit_access(PrivilegeCodes.CHANGE))
         else:
             self.assertIn(self.mike, group.gaccess.get_users_with_explicit_access(PrivilegeCodes.OWNER))
-
-    def _set_request_message_attributes(self, request):
-        # the following 3 lines are for preventing error in unit test due to the view being tested uses
-        # messaging middleware
-        setattr(request, 'session', 'session')
-        messages = FallbackStorage(request)
-        setattr(request, '_messages', messages)
